+++ .qtenvrc
@@ -12,4 +12,4 @@ run-max-animation-speed=nan
 run-max-playback-speed=100
 run-min-animation-speed=nan
 run-min-playback-speed=0.01
-run-playback-speed=0.3019951720402016
+run-playback-speed=0.24547089156850305
